requirements.txt
ipython
# exercise 9
pytest
netmiko
pyyaml
# pytest-clarity # ImportError: libffi.so.6: cannot open shared object file: No such file or directory
# exercise 11
graphviz
# exercise 12
tabulate
# exercies 17
pyyaml
# exercise 20
jinja2
